Attempt OOP #2
I was able to fix an attribute error that gave me a massive headache. The reason why it was appearing was because I initialized the list incorrectly. I was using just one [] instead of 3 for each variable, ip,port & city. This also made the code look more beautiful and clean compared to the original version.