Implemented a new Class & bold characters
The startProgram() class will execute the program within a loop to allow the user to enter multiple IPs if he/she wants. I also added an exception for exiting the program, this will print a exit message.I also added bold text to the user input and exception, to make the program more pretty.